+++ model/src/main/java/agency/amazon/test/model/TrafficByAsin.java
@@ -2,6 +2,8 @@ package agency.amazon.test.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TrafficByAsin {
     private int browserSessions;
@@ -10,26 +12,26 @@ public class TrafficByAsin {
     private int mobileAppSessionsB2B;
     private int sessions;
     private int sessionsB2B;
-    private double browserSessionPercentage;
-    private double browserSessionPercentageB2B;
-    private double mobileAppSessionPercentage;
-    private double mobileAppSessionPercentageB2B;
-    private double sessionPercentage;
-    private double sessionPercentageB2B;
+    private BigDecimal browserSessionPercentage;
+    private BigDecimal browserSessionPercentageB2B;
+    private BigDecimal mobileAppSessionPercentage;
+    private BigDecimal mobileAppSessionPercentageB2B;
+    private BigDecimal sessionPercentage;
+    private BigDecimal sessionPercentageB2B;
     private int browserPageViews;
     private int browserPageViewsB2B;
     private int mobileAppPageViews;
     private int mobileAppPageViewsB2B;
     private int pageViews;
     private int pageViewsB2B;
-    private double browserPageViewsPercentage;
-    private double browserPageViewsPercentageB2B;
-    private double mobileAppPageViewsPercentage;
-    private double mobileAppPageViewsPercentageB2B;
-    private double pageViewsPercentage;
-    private double pageViewsPercentageB2B;
-    private double buyBoxPercentage;
-    private double buyBoxPercentageB2B;
-    private double unitSessionPercentage;
-    private double unitSessionPercentageB2B;
+    private BigDecimal browserPageViewsPercentage;
+    private BigDecimal browserPageViewsPercentageB2B;
+    private BigDecimal mobileAppPageViewsPercentage;
+    private BigDecimal mobileAppPageViewsPercentageB2B;
+    private BigDecimal pageViewsPercentage;
+    private BigDecimal pageViewsPercentageB2B;
+    private BigDecimal buyBoxPercentage;
+    private BigDecimal buyBoxPercentageB2B;
+    private BigDecimal unitSessionPercentage;
+    private BigDecimal unitSessionPercentageB2B;
 }
